requirements.txt
streamlit
numpy
matplotlib
sentence-transformers
faiss-cpu